Fixed error in displaying of name of genera
see: https://jira.naturalis.nl/browse/NSRLNG-185
Parents name was blindly removed from the child taxa's names in the
taxonomic tree branch in the left column. This caused "Adinetidae" with
parent "Adinetida" to be reduced to just "e". Replacement is now done
only for taxa of the rank species and lower.